requirements.txt
Django==1.11.0
Padding==0.4
SQLAlchemy==1.3.1
dj-database-url==0.3.0
dj-stripe==0.3.5
django-braces==1.4.0
django-clear-cache==0.3
django-cors-headers==3.0.1
django-heroku-memcacheify==0.8
django-jsonfield==0.9.13
django-keen==0.1.3
django-markdown-deux==1.0.5
django-model-utils==2.2
django-postgrespool==0.3.0
django-pylibmc==0.5.0
djangorestframework==3.9.4
drf-json-api==0.1.0
gunicorn==19.9.0
keen==0.3.7
markdown2==2.3.0
psycopg2==2.5.4
pycrypto==2.6.1
pylibmc==1.6.0
requests==2.5.1
six==1.8.0
stripe==1.19.1
whitenoise==4.1.2
wsgiref==0.1.2
